PLANET-5189 ImageArchive: separate render props
* Move render props to a separate function that takes both the comopnent
that's passing the render prop and the component that gets the render
prop. This reflects the render prop function's dependency on both. This
way the render prop function can easily use the API of both components.
